mount_null(8) man page is installed to /usr/local/share/man.
mount.nullfs -> mount_null to match the man page name.

code signing in an attempt to defeat Apple's
`com.apple.private.nullfs_allow`, however this did not work -
as the entitlement is _private_.

Have to re-implement it as bindfs, by renaming all the variables and
functions into a separate bindfs module BUT WITHOUT the entitlement check.